requirements.txt
praw
pandas
dnspython
pymongo